Further Organization & Some Refactoring
Moved Lobby Methods from Utility to Lobby.cs.
Added multiple regions
Moved order of multiple methods in ItemManagement and DrawMenu
DrawESPMenu -> DrawRenderMenu
